services/product-finder/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
fastapi==0.117.1
frozenlist==1.7.0
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==4.9.3
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
selenium==4.18.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
webdriver-manager==4.0.1
websockets==15.0.1
yarl==1.20.1
